code/requirements.txt
numpy
pandas
gensim
lasagne
progressbar2
deepdish
scikit-learn
multiprocessing
mrec